Update test_accounts.py
I added the additional fields first name, last name and skills, as well as addressed some of the things that Sam had brought up in their review of my code. I wasn't aware how username was handled before, so I tried to fix that. The only thing I didn't address was the call to delete an account in the database, because I don't know how you would simulate deleting an account without actually deleting one. I know this accesses the database, but I don't know how to work around that